Re-arranges tests
